initial commit
This commit starts the git history of this project which should be an
example of clean Android development using the Model View Presenter
(MVP) architecture.

For now, this app pulls down Github users from the Github API and
displays them in a list. If you scroll to the bottom of the list, you
will see a progress bar indicating that a new batch of users is loading
and the results are displayed when ready.