fix recycler view not displaying textview properly
This commit fixes a bug where after loading in new users into the
list, the text views and divider views were not showing up unless you
scrolled around and came back to them. This also caused the progress bar
item to not display full width on first load.

Originally, this bug was thought to be a problem with the emulator.