implement local storage
This commit adds Realm to the project, implements the local data source,
adds a few necessary methods for saving and deleting users in storage,
adds some tests for only local storage, and finally corrects some issues
in other unit tests due to Robolectric not playing well with Realm and
its use of Loopers.

Also, we add in the RxBinding library for RecyclerView in order to
more cleanly detect when to start loading more users in the list.

I'm not proud of the size of this commit...