add swipe refresh, refactor dependency injection
This commit adds code to enable a swipe to refresh layout on the
recycler view. This change requires new contract interfaces, which means
more tests.

Also, this commit refactors the dependency injection. We no longer infer
objects that should be injected into each method. There is now a
injection utils class that performs common injection calls which will be
needed if we were to add new repositories or api endpoints.